Initial import of GI-DocGen
GI-DocGen is a tool that takes introspection data and ancillary MarkDown
files, and turns them into an API reference for a GObject-based library.

Ideally, it will replace gtk-doc. In practice, it's currently an
experiment in how much we can stretch introspection data and MarkDown
until either, or both, of them fall apart.