generate: Add a fully qualified type name field
Types within the same namespace only have a relative name; we should
build a fully qualified type name (FQTN) that matches the type name of
types coming from dependencies. In the same way, we should split up the
namespace and type name of types coming from dependencies, so we have
the same behaviour.

This way, we can drop a bunch of string concatenation in the templates,
and we can also have a way to do cross-linking later.