generate: Split namespace from type name
Symbols coming from the same namespace have a name relative to the
namespace; symbols coming from dependencies, on the other hand, have a
fully qualified name.

We introduce the concept of a fully qualified type name (fqtn) for every
type we have, and add separate namespace and name attributes alongside
it. This allows us to refer to a type via its fqtn, and later on it'll
allow us to resolve a type name through its namespace.