gir: Restructure the type resolution
Instead of naively dumping the types that we find, let's try to work a
bit harder to resolve all the types.

First of all, we store all the ast.Type instances into a global LUT, so
we return the same instance for the same fully qualified type name. We
do this for every namespace we end up parsing.

Additionally, once we finished parsing the main namespace, we try and
resolve all the C types and class ancestors, and replace everything that
takes a string with a real type.

This should give the template code a more complete view of what it has,
and we are going to be able to resolve things like links more easily.