Add a Markdown extension for internal links
We use an inline processor to match a pattern like:

    [<fragment>@<endpoint>]

where "fragment" identifies the kind of type we are linking to, and
"endpoint" is a type name, a property name, a signal name, or any other
fully qualified end point. Once we match the pattern, we write a link
element and inject it into the HTML tree.

The extension is kind of crude, at the moment; it needs configuration to
deal with identifier and symbol prefixes, the namespace in use, as well
as adding CSS classes and other niceties to the generated link.

This extension is only meant to work with intra-namespace links; links
to other dependencies are going to require more finesse, and an
additional transformation to take into account the relative path of each
dependency's documentation.