Parse links in the preprocessing stage
The Markdown extension route is nice, but passing arguments to it is
less than ideal, and gets complicated fast.

We can achieve pretty much the same in the preprocessing step we still
need to run, so might as well do that there.

This allows us to parse links and replace the text with the various
identifier and symbol prefixes coming straight from the namespace data.