gir: Handle types disguised as `void*`
A pattern in our C APIs it to us untyped generic pointers in place of
typed ones to avoid excessive casting. We usually specify the actual
type in the annotations for the function. This leads to a type element
in the introspection data like:

    <type name="Widget" c:type="gpointer"/>

which then gets registered in our type LUT as:

    Type(name="Gtk.Widget", ctype="gpointer")

So, when we look up a type using just its FQTN, we might end up with the
incomplete type object, and it confuses things when we try to get the C
type out of it later on.

We should skip registering disguised gpointer/void* types.

Fixes: #32