generate: Remove unnecessary duplication
We shouldn't repeat ourselves this much when it comes to generating the
dummy index dictionaries we pass to the templates.

Cutting down the amount of work reduces the time to generate the API
reference as well; even with the newly added recursion through the
ancestors and interfaces, a full build for Gtk-4.0 goes back from 75
seconds to 50 seconds.