generate: Handle known attributes
Attributes are custom key/value annotations that are reported "as is" in
the XML. We can define a keys that library authors can use in their
annotations, and when we find them in the GIR, we can use them to
generate additional content.

For instance, if a property has an "org.gtk.Property.get" annotation
pointing to a getter function, we can generate the link in the property
documentation.