generate: Improve type caching and resolution
We should always have a C type associated to a type, by the time we
close the ast.Repository object. This requires multiple passes to clean
up ast.Type() instances that have no c:type associated to them; it also
requires storing the namespace of each type when we parse it, so we can
always determine where a type comes from.