gir: Handle empty type nodes
Empty type nodes are not really expected to be introspectablel but since
we are a documentation generator we are also not looking at the
"introspectable" field.

There's only one case in GTK where the `<type>` element is completely
empty: the `<property>` element representing the GtkPrinter:backend
property. The GtkPrintBackend type is private to GTK, but the C type is
visible as a record; of course, this is not enough for a public
property: the GType *must* be available.

We can cope with this wart in the introspection data by simply mark the
type as "void".