template: Hook up the "go to top" button v2
Show the button when the content scrolls for more than an arbitrary
amount of pixels; when pressing it, we scroll back to the top. This
should help with long pages.

The button is above the content block, to avoid contributing to the
sizing and layout, since it'll be floating at a fixed position above
the actual content.