Re-implement the link generator
This is a larger commit than I intended to, but the link generator
touches a lot more than just the text processing side of things.

First of all, we need to take into account things like the moved-to
attribute, which requires removing callables that have been moved from
the global namespace to a specific type.

Additionally, we need to store properties and signals inside maps, to
avoid excessive iteration on something that we know is unique per type.

Finally, we cannot just split the link tokens and call it a day: we
should use pattern matching, wherever appropriate, and validate every
token we get against the introspection data. This allows us to catch
typos and bad links, but more importantly: it allows us to implement
support for global functions, type functions, virtual functions, class
methods, and methods in a more consistent way, with fallbacks in place
for short-hand notations.