utils: Generate external reference links
Links to internal data types and symbols are easy to craft. External
links to other references are, on the other hand, a bit more complicated
because the link generator has no idea where those namespaces are.

Instead of complicating the link generator any further, we punt the
problem of resolving external links to other API references created by
gi-docgen by creating specially crafted links; these external links
have:

 - an "external" CSS class
 - a `data-namespace` data attribute, with the namespace of the symbol
 - a `data-link` data attribute, with the link to the symbol
 - an empty, `javascript:void(0)` href attribute

This means that the links point to nowhere by default, but they can be
parsed when loading the page, and the namespace mapped to the desired
URL.