generate: Add external links remapping
Links to external namespaces can be rewritten using an URL map, a simple
association between namespace name and the base URL.

In the basic template we inject a bit of JavaScript that rewrites the
href attribute of every external link using the URL map provided by the
project configuration whenever the page is loaded.

This should also allow rewriting links when uploading the API reference
to a web server, by simply replacing the URL map file.