genindices: Rewrite the index generator
We only generate a JSON file, now, but it contains all the types,
methods, functions, properties, and signals defined in the namespace.
Additionally, the index contains all the unique terms present in the
documentation, pointing to all the objects that reference them.

This new index can be used to implement a simple term based search.