Return to setuptools
Poetry, as nice as it is, with its declarative approach to the build, is
also really meant to be used with Python modules, not Python CLI tools,
as it lacks the ability to install files outside the module's directory.

Let's go back to the venerable setuptools, minus the bugs we had.