Add pkg-config file for gi-docgen
Sadly, a lowest common denominator solution.

We need to override the build command for setuptools to generate a
pkg-config file from a template.

To avoid having different results with the Meson build, we cannot use
the pkgconfig.generate() function, and instead rely on configure_file().

Such is the life of maintainers of projects using multiple build
systems.

Fixes: #60