template: Search through the content files
Now that the index has a "content" type for the documents, we need to
handle it properly.

Since we show the title of the content files in the results, we need to
move the "<code>…</code>" elements from the result generator to a
separate function, so that we don't show document titles in the same way
as identifiers and symbols.