template: Switch to Solarized for syntax highlighting
The default theme doesn't really work well in dark mode, and even in
light mode it "screams" a lot.

The Solarized color scheme, on the other hand, looks better in light
mode, and has a dark mode as well.

We can use the @import rule to determine whether to load the light or
dark Solarized scheme depending on the browser settings.

Fixes: #101