Add tooltips to class hierarchy edges
Currently, when hovering over an edge between a class and its ancestor,
it gives a tooltip similar to `this--ancestor_0`.

This is not really informative and is probably even worse for people
with screen readers, as they can't see the diagram and can only get
confused by it.

Replace it with `$TYPE extends $TYPE` or `$TYPE implements $TYPE`.