enum template: Add more titles to navigation bar
The current enum template only mentions the "Description" title (and
"Functions" if applicable). However, depending on how long the
description might be, it's a good idea to also add a link for quick
access to the members. There's also an error domain below that, so let's
link that as well.