generate: Use const.*.html instead of constant.*.html for search index links
HTML files for constants are generated like const.*.html, but the keywords index
links constants keywords with a constant.*.html file, although the documentation
outline itself uses the correct const.*.html file. This leads to constants being
properly accessible while browsing through the outline tree, but not accessible
when using the search result coming from the index, which leads to a No such file
or directory error.

So this commit fixes it by changing constant.*.html to the correct const.*.html
for the keywords index.