indexing: Don't generate 'terms' in JSON index
Since the 'symbols' list in the JSON index is now being sorted, the
row-based references in the 'terms' list are invalidated. As such,
there's no reason to generate them at all.

Do away with all index-based symbol lookups, and all parsing and
stemming previously used to generate the "terms" list.