generate: Use the appropriate namespace for indices
When including the ancestors and implemented interfaces of a class in
its index, we need to resolve eventual links in the short documentation
using the appropriate namespace, not the current one.

This allows us to fix the case in which an ancestor in another namespace
uses a link without specifying the namespace.

Fixes: #137