template: Use definition lists for callable arguments
Tables are hard to size, and they get very wonky on mobile, or narrow
window widths.

Both MSDN and Apple's developer docs use definition lists for parameters
to a callable, so might as well follow them.

We can also drop the table for the return value entirely, since the
return value does not have a name.

Fixes: #123