setuptools: simplify getting metadata via static config
A couple things were in both setup.cfg but with the wrong contents, and
in setup.py with the correct contents -- if both files specify a
metadata entry, then the former is always ignored and overridden by the
latter

- long-description
- license
- console_scripts

Using setup.cfg has the unique property which you cannot do with
setup.py, of automatically reading in a file as a metadata value --
using the declarative cfg for this, allows dropping a custom function to
read a file. This can be used for both long-description and version.

The console_scripts in the cfg said to name the entry point script
"meson", which is obviously wrong and would make it impossible to run
gi-docgen if not for the setup.py override. Clean up that inconsistency.