Remove unnecessary dummy_dep
meson.override_find_program() has existed since 0.46. In order to use
gi-docgen as a subproject now, use the following in a wrap file.

```ini
[provide]
program_names=gi-docgen
```

Then a consumer of gi-docgen just needs to use:

```meson
gi_docgen = find_program('gi-docgen')
```

If gi-docgen is not found on the system, then the gi-docgen wrap will be
triggered.

Note that it was already possible to use gi-docgen in this way, but this
should push consumers into better Meson patterns.