Redesign the search results
The table approach does not work on mobile, or smaller windows. Let's
pull a search engine, and show the results in a definition list.

While we're at it, let's use a similar approach as the various indices,
and add emblems for the result type and whether or not a symbol is
deprecated.