generate: Simplify section generation for compound types
Instead of manually going through the various sections, we can avoid
repeating ourselves and iterate over an array of dictionaries that
define each section.

Since we pass the indices to the templates, we can also avoid manually
describing the structure of the side bar.

This change should reduce the amount of custom code and templating that
goes into the indices.