generate: Add origin metadata field for content files
The "origin" metadata field can be used to point to the name of the file
in the repository.

For instance, if a content file is generated from a template into
another directory, the generated file can still refer to the template
file.

The "origin" field is used to generate the location of the source file
in the source repository.

If the "origin" field is not present, it defaults to the base name of
the content file.