generate: Add earliest_version option for since tags
This improves the strategy introduced in !198. By default, if a Since
tag is not present on an API, we assume the version to use matches the
API version. But this doesn't work for projects where API version
doesn't match the project major version, so add a new earliest_version
option to allow overriding this heuristic.

Fixes #179