Add mermaid.js support to the Markdown parser
Use a local Markdown extension that identifies code blocks using the
Mermaid identifier, and expands them to a <pre>...</pre> block without
code highlighting, so that mermaid.js can do its job.

We include the minified version of mermaid.js, as it's the smallest,
self-contained one; the non-minified version clocks at 7 MB, and the ESM
version is just a shim for a bunch of separate modules.

Fixes: #180