Revert "generate: Add fallback for missing "since""
This reverts commit 4fa286d504df8dd970a175db806a7f9156cdf527.

Turns out that a couple libraries (*cough* WebKitGTK *cough*) don't
match the namespace version with their library version, so the fallback
and the availability become absolute nonsense.

The proper fix is to have the template handle the fallback using the
configuration to determine the base version that introduced a symbol. Or
those libraries stop trying to be too clever for their own good.

Fixes: #179