utils: Append period after last line in more cases
After the last line in a block, gi-docgen appends a period. Obviously,
this should not happen when there is already a period. This is handled
by testing for isalpha() on the last character.

However, periods should also be inserted after links or inline code. Fix
this by broadening the cases in which a period is appended.

Fixes #181