Support static virtual methods
GObject-introspection now supports generating GIR data for "static"
virtual methods: functions that do not have an instance parameter, and
are typically used to construct new instances, like
GObjectClass.constructor().

We currently rely on the existance of an instance parameter element in
order to generate virtual function templates, so we are going to need to
check for its presence before trying to unpack it.

See: GNOME/gobject-introspection!361

Fixes: #211