Tests: Fix existing test_gtkdoc_sigils()
The tests in this function were all written using `assertTrue()`
instead of `assertEqual()`, which means they weren't testing
anything.

The second argument of `self.assertTrue()` is a _message_ to attach
to the test, it isn't compared to the first argument. (That's what
`assertEqual()` is for.) So, for a test like this:

```python
self.assertTrue(
    mdext.process_gtkdoc_sigils("will emit the #GCancellable::cancelled signal."),
    "will emit the `GCancellable::cancelled` signal.")
```

Literally anything could appear in that second argument, and the test
would still pass as long as the first argument evaluates `True`.

```python
    # Still passes!
self.assertTrue(
    mdext.process_gtkdoc_sigils("will emit the #GCancellable::cancelled signal."),
    "Bleep bloop I'm a message")

    # Actually fails
self.assertEqual(
    mdext.process_gtkdoc_sigils("will emit the #GCancellable::cancelled signal."),
    "Now I'm a value to be compared to the previous arg")
```