Use dot's "svg_inline" output format
`utils.render_dot` doesn't have to do a lot of processing of the
SVG output to remove the svg-file headers, if the SVG is generated
with `dot -Tsvg_inline` instead of `dot -Tsvg`.

Also, add a basic functionality test for `utils.render_dot` to
`test_utils.py`. (The test also checks that the SVG generated
is clear of `<?xml` or `DOCTYPE` strings.)