gidocgen: expirimental markdown theme
This starts on an expirimental markdown theme that would allow writing
documentation as a series of Markdown files. This could be advantageous
for a couple reasons.

First, it would allow tooling like Foundry to be able to lookup
documentation in a format that does not require a full webbrowser. That
would provide flexibility in terms of where we can show those docs.
Whether that is part of a completion window, tooltips, or more.

Additionally, it could be used by language models to prime context more
easily so that API use is digestable without an extremely large context
window. This could also be automated by foundry with the documentation
subsystem for quick recall on how a function works.