Make sure to escape query strings
Unescaped query strings should not be passed to the HTML parser, to
avoid unwanted execution of JavaScript.

The query is shown in the header of the search results, so we can easily
split the header from the results; then we use a plain text node to
represent the query, and let the browser escape it.

See: https://cheatsheetseries.owasp.org/cheatsheets/DOM_based_XSS_Prevention_Cheat_Sheet.html

Fixes: #228