Override licensing for template files
The template files include templated REUSE tags so that the generated
files can pass REUSE linting. This trips up the reuse linter, though, so
we need to ensure that the licensing and copyright notices are taken
from an ancillary file to avoid an error.

The REUSE.toml file used by newer versions of the REUSE spec allows us
to do that, so we need to convert the existing (and deprecated) dep5
file to the new REUSE.toml one.